added a buildable Xcode project. crashed on run due to 32 vs 64 bit issues

+++ include/objects.h
@@ -50,7 +50,7 @@ const Value V_NIL = IMMED_V(IMMED_SPECIAL, 0);  // a.k.a. 2
 
 const Value V_TRUE = IMMED_V(IMMED_BOOLEAN, 1); // a.k.a. 0x1A
 
-inline int      V_TAG(Value v)      { return ((int) v) & TAG_MASK; }
+inline int      V_TAG(Value v)      { return ((int)(size_t) v) & TAG_MASK; }
 
 /// @defgroup preds Type predicates
 /// Tests for Value types.
@@ -66,7 +66,7 @@ inline bool     V_ISPTR(Value v)    { return (V_TAG(v) & TAG_PTR); }
 
 /// Is the Value a character?
 
-inline bool     V_ISCHAR(Value v)   { return (((int) v) & (TAG_MASK | IMMED_MASK)) == (TAG_IMMED | IMMED_CHAR); }
+inline bool     V_ISCHAR(Value v)   { return (((int)(size_t) v) & (TAG_MASK | IMMED_MASK)) == (TAG_IMMED | IMMED_CHAR); }
 
 /// Is the Value a reference to a Binary object?
 
@@ -127,15 +127,15 @@ EXPORT int V_CHAR_error(void);
 /// first -- use only if you're sure the Value is an int (or it doesn't much
 /// matter if it isn't).
 
-inline int      UNSAFE_V_INT(Value v)   { return (((int) v) >> 2); }
+inline int      UNSAFE_V_INT(Value v)   { return (((int)(size_t) v) >> 2); }
 
 /// Converts a Value to an int. Throws if Value is not an integer.
 
-inline int      V_INT(Value v)          { return V_ISINT(v) ? (((int) v) >> 2) : V_INT_error(); }
+inline int      V_INT(Value v)          { return V_ISINT(v) ? (((int)(size_t) v) >> 2) : V_INT_error(); }
 
 /// Converts a Value to a Unicode character. Throws if Value is not a character.
 
-inline int      V_CHAR(Value v)         { return V_ISCHAR(v) ? (((int) v) >> 4) : V_INT_error(); }
+inline int      V_CHAR(Value v)         { return V_ISCHAR(v) ? (((int)(size_t) v) >> 4) : V_INT_error(); }
 
 /// Converts a Value to a bool. Result is false if Value is V_NIL, otherwise true.
