+++ include/predefined.h
@@ -11,7 +11,7 @@
 #define __PREDEFINED_H__
 
 #define PREDEF_NAME(name)   predef_##name
-#define V_PTRTO(name)       ((Value) (((int) &name) + 1))
+#define V_PTRTO(name)       ((Value) (((int)(size_t) &name) + 1))
 #ifdef _MSC_VER
   #ifdef NSRT_BUILD
    #define DLLSPEC __declspec(dllexport)
